Enabled local timezone advertisement
This feature is specified in 802.11v-2011 7.3.2.87.

The timezone is supposed to be specified via beacon tag 98 (0x62) as a
string in "POSIX format"
(https://developer.ibm.com/articles/au-aix-posix/#understanding-the-posix-format3).

Happily, the POSIX format timezone is easily extracted from the tzfile
format (https://stackoverflow.com/a/67045287), and /etc/localtime is a
symlink to such a file.

I'm not sure what OSes, if any, will detect these tags and use them to set
timezone.  Most likely it would only be used as a fallback from some
"location-based" timezone determination algorithm.

- Android 5.x source alludes to knowledge of this 802.11 feature in `DateTimeSettings.java`:
  https://android.googlesource.com/platform/packages/apps/Settings/+/android-5.0.1_r1/src/com/android/settings/DateTimeSettings.java#113
- … but more recent versions of Android don't appear to mention it at all:
  https://android.googlesource.com/platform/packages/apps/Settings/+/main/src/com/android/settings/datetime/DateTimeSettings.java

FIXME: The capability of advertising local timezoney was added to hostapd in
https://chromium.googlesource.com/external/w1.fi/cgit/hostap/+/39b97072b2a45551e6f20e6251eeaca269f22a2d%5E%21/#F1
with the 'time_zone' variable, but appears to be broken in hostapd v2.9.